# Estimators speed improvements.