--- a/docs/articles/using_cvCovEst.html
+++ b/docs/articles/using_cvCovEst.html
@@ -31,7 +31,7 @@
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>
 
@@ -88,7 +88,7 @@
 <a href="https://pboileau.ca">Philippe Boileau</a>, <a href="https://github.com/bcollica">Brian Collica</a>, <a href="https://nimahejazi.org">Nima Hejazi</a>
 </h4>
             
-            <h4 class="date">2021-07-24</h4>
+            <h4 class="date">2021-09-07</h4>
       
       <small class="dont-index">Source: <a href="https://github.com/PhilBoileau/cvCovEst/blob/master/vignettes/using_cvCovEst.Rmd"><code>vignettes/using_cvCovEst.Rmd</code></a></small>
       <div class="hidden name"><code>using_cvCovEst.Rmd</code></div>
@@ -113,7 +113,7 @@
 <div class="sourceCode" id="cb1"><pre class="downlit sourceCode r">
 <code class="sourceCode R"><span class="kw"><a href="https://rdrr.io/r/base/library.html">library</a></span><span class="op">(</span><span class="va"><a href="http://www.stats.ox.ac.uk/pub/MASS4/">MASS</a></span><span class="op">)</span>
 <span class="kw"><a href="https://rdrr.io/r/base/library.html">library</a></span><span class="op">(</span><span class="va"><a href="https://github.com/PhilBoileau/cvCovEst">cvCovEst</a></span><span class="op">)</span></code></pre></div>
-<pre><code>## cvCovEst v1.0.0: Cross-Validated Covariance Matrix Estimation</code></pre>
+<pre><code>## cvCovEst v1.0.1: Cross-Validated Covariance Matrix Estimation</code></pre>
 <div class="sourceCode" id="cb3"><pre class="downlit sourceCode r">
 <code class="sourceCode R"><span class="fu"><a href="https://rdrr.io/r/base/Random.html">set.seed</a></span><span class="op">(</span><span class="fl">1584</span><span class="op">)</span>