--- a/docs/news/index.html
+++ b/docs/news/index.html
@@ -71,7 +71,7 @@
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>
 
@@ -126,9 +126,18 @@
       <small>Source: <a href='https://github.com/PhilBoileau/cvCovEst/blob/master/NEWS.md'><code>NEWS.md</code></a></small>
     </div>
 
+    <div id="cvcovest-101-2021-09-07" class="section level1">
+<h1 class="page-header" data-toc-text="1.0.1">
+<a href="#cvcovest-101-2021-09-07" class="anchor"></a>cvCovEst 1.0.1 (2021-09-07)<small> Unreleased </small>
+</h1>
+<ul>
+<li>
+<code><a href="../reference/scadEst.html">scadEst()</a></code> and <code><a href="../reference/adaptiveLassoEst.html">adaptiveLassoEst()</a></code> are now vectorized, greatly improving their computational efficiency. Thanks, Brian!</li>
+</ul>
+</div>
     <div id="cvcovest-100-2021-07-24" class="section level1">
 <h1 class="page-header" data-toc-text="1.0.0">
-<a href="#cvcovest-100-2021-07-24" class="anchor"></a>cvCovEst 1.0.0 (2021-07-24)<small> Unreleased </small>
+<a href="#cvcovest-100-2021-07-24" class="anchor"></a>cvCovEst 1.0.0 (2021-07-24)<small> 2021-07-25 </small>
 </h1>
 <ul>
 <li>