--- a/docs/pkgdown.yml
+++ b/docs/pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   using_cvCovEst: using_cvCovEst.html
-last_built: 2021-07-24T18:08Z
+last_built: 2021-09-07T12:21Z