--- a/docs/reference/adaptiveLassoEst.html
+++ b/docs/reference/adaptiveLassoEst.html
@@ -81,7 +81,7 @@ resulting in larger values being penalized less and reducing bias
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>