--- a/docs/reference/adaptiveLassoThreshold.html
+++ b/docs/reference/adaptiveLassoThreshold.html
@@ -74,7 +74,7 @@ is meant to be applied to sample covariance matrix" />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>