--- a/docs/reference/bandingEst.html
+++ b/docs/reference/bandingEst.html
@@ -78,7 +78,7 @@ Bickel and Levina (2008)
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>