--- a/docs/reference/bestInClass.html
+++ b/docs/reference/bestInClass.html
@@ -74,7 +74,7 @@ finds the associated hyper-parameters if applicable." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>