--- a/docs/reference/checkArgs.html
+++ b/docs/reference/checkArgs.html
@@ -73,7 +73,7 @@ passed to cvCovEst() function meet its specifications." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>