--- a/docs/reference/checkPlotSumArgs.html
+++ b/docs/reference/checkPlotSumArgs.html
@@ -75,7 +75,7 @@ additional arguments may be checked at the individual function level." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>