--- a/docs/reference/cvCovEst.html
+++ b/docs/reference/cvCovEst.html
@@ -73,7 +73,7 @@ estimator from among a set of candidate estimators." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>
 
@@ -304,7 +304,7 @@ estimators' loss over the folds of the cross-validated procedure.</p></li>
 #&gt; $args$cv_loss
 #&gt; <span style='font-weight: bold;'>&lt;quosure&gt;</span>
 #&gt; expr: <span style='color: #0000BB;'>^cvMatrixFrobeniusLoss</span>
-#&gt; env:  <span style='color: #0000BB;'>0x7fa3ef508758</span>
+#&gt; env:  <span style='color: #0000BB;'>0x7f9bfc30a138</span>
 #&gt; 
 #&gt; $args$cv_scheme
 #&gt; [1] "v_fold"