--- a/docs/reference/cvFrobeniusLoss.html
+++ b/docs/reference/cvFrobeniusLoss.html
@@ -75,7 +75,7 @@ over a fold object (from 'origami'
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>