--- a/docs/reference/cvMatrixFrobeniusLoss.html
+++ b/docs/reference/cvMatrixFrobeniusLoss.html
@@ -77,7 +77,7 @@ selections, but is more computationally efficient." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>