--- a/docs/reference/cvMultiMelt.html
+++ b/docs/reference/cvMultiMelt.html
@@ -74,7 +74,7 @@ map corresponds to a different estimator." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>