--- a/docs/reference/cvRiskByClass.html
+++ b/docs/reference/cvRiskByClass.html
@@ -75,7 +75,7 @@ and maximum.  The results are output as a tibble." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>