--- a/docs/reference/cvScaledMatrixFrobeniusLoss.html
+++ b/docs/reference/cvScaledMatrixFrobeniusLoss.html
@@ -80,7 +80,7 @@ are of different magnitudes." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>