--- a/docs/reference/cvSummaryPlot.html
+++ b/docs/reference/cvSummaryPlot.html
@@ -76,7 +76,7 @@ class is also provided if applicable." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>