--- a/docs/reference/denseLinearShrinkEst.html
+++ b/docs/reference/denseLinearShrinkEst.html
@@ -81,7 +81,7 @@ Schäfer and Strimmer (2005)
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>