--- a/docs/reference/getHypers.html
+++ b/docs/reference/getHypers.html
@@ -73,7 +73,7 @@ hyperparameters associated with an estimator passed to cvCovEst()." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>