--- a/docs/reference/linearShrinkEst.html
+++ b/docs/reference/linearShrinkEst.html
@@ -78,7 +78,7 @@ values near 0 are more likely to be be very biased but have low variance." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>