--- a/docs/reference/linearShrinkLWEst.html
+++ b/docs/reference/linearShrinkLWEst.html
@@ -79,7 +79,7 @@ by Ledoit and Wolf (2004)
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>