--- a/docs/reference/multiHyperRisk.html
+++ b/docs/reference/multiHyperRisk.html
@@ -78,7 +78,7 @@ plot is not generated for that estimator class." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>