--- a/docs/reference/plot.cvCovEst.html
+++ b/docs/reference/plot.cvCovEst.html
@@ -75,7 +75,7 @@ estimator using cvCovEst." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>