--- a/docs/reference/plotRobustPoetEst.html
+++ b/docs/reference/plotRobustPoetEst.html
@@ -73,7 +73,7 @@ the cross-validated risk of the Robust POET estimator." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>