--- a/docs/reference/robustPoetEst.html
+++ b/docs/reference/robustPoetEst.html
@@ -89,7 +89,7 @@ resulting estimate are not guaranteed to be equal to one." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>