--- a/docs/reference/sampleCovEst.html
+++ b/docs/reference/sampleCovEst.html
@@ -73,7 +73,7 @@ function is a simple wrapper around covar()." />
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>