--- a/docs/reference/scadEst.html
+++ b/docs/reference/scadEst.html
@@ -80,7 +80,7 @@ and hard thresholding after 3.7 * lambda
       </button>
       <span class="navbar-brand">
         <a class="navbar-link" href="../index.html">cvCovEst</a>
-        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.0</span>
+        <span class="version label label-default" data-toggle="tooltip" data-placement="bottom" title="Released version">1.0.1</span>
       </span>
     </div>