… docs/reference/scadThreshold.html
… docs/reference/strToNumber.html
… docs/reference/summary.cvCovEst.html
… docs/reference/taperingEst.html
… docs/reference/theme_cvCovEst.html
… docs/reference/thresholdingEst.html