This PR includes changes to the scadEst and adaptiveLassoEst functions to boost speed performance.  A vectorized version of the scadThreshold function is also implemented.  The new functions were benchmarked against the old ones using a random matrix of dimension n=100, p=100.  Runtime for the adaptive Lasso was reduced from 54ms to 3ms and the SCAD went from 218ms to 11.8ms.  

![Screen Shot 2021-09-06 at 10 49 52 AM](https://user-images.githubusercontent.com/59424341/132251279-b5423ec8-eef2-4417-89a4-5947efc0446d.png)